requirements.txt
langchain>=0.1.20
langchain-community>=0.0.45
langchain-huggingface==0.1.2
groq==0.3.0
python-dotenv==1.0.1
faiss-cpu==1.7.4



